interface/requirements.txt
pyautogui
pillow
opencv-python
grpcio==1.39.0
grpcio-tools==1.39.0
